refactor: Timetable 분리
